requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.116.0
h11==0.16.0
idna==3.10
lxml==6.0.0
outcome==1.3.0.post0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.4
requests-toolbelt==1.0.0
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.46.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
apscheduler==3.10.4
psycopg2-binary==2.9.9
aiohttp==3.10.10
